Move search dto files from persistence to domain, breaking the dependency from client to persistence.
