+++ mauro-api/src/test/groovy/uk/ac/ox/softeng/mauro/search/SearchIntegrationSpec.groovy
@@ -1,6 +1,5 @@
 package uk.ac.ox.softeng.mauro.search
 
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchRequestDTO
 import uk.ac.ox.softeng.mauro.testing.CommonDataSpec
 
 import jakarta.inject.Inject
@@ -10,7 +9,8 @@ import uk.ac.ox.softeng.mauro.domain.datamodel.DataModel
 import uk.ac.ox.softeng.mauro.domain.folder.Folder
 import uk.ac.ox.softeng.mauro.persistence.ContainerizedTest
 import uk.ac.ox.softeng.mauro.persistence.datamodel.DataModelContentRepository
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchResultsDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchRequestDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchResultsDTO
 import uk.ac.ox.softeng.mauro.web.ListResponse
 
 @ContainerizedTest
