+++ mauro-api/src/test/groovy/uk/ac/ox/softeng/mauro/security/AdminAccessIntegrationSpec.groovy
@@ -1,6 +1,6 @@
 package uk.ac.ox.softeng.mauro.security
 
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchRequestDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchRequestDTO
 
 import io.micronaut.http.HttpStatus
 import io.micronaut.http.client.exceptions.HttpClientResponseException
@@ -11,7 +11,7 @@ import spock.lang.Shared
 import uk.ac.ox.softeng.mauro.domain.datamodel.DataModel
 import uk.ac.ox.softeng.mauro.domain.folder.Folder
 import uk.ac.ox.softeng.mauro.persistence.SecuredContainerizedTest
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchResultsDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchResultsDTO
 import uk.ac.ox.softeng.mauro.web.ListResponse
 
 @SecuredContainerizedTest
