+++ mauro-client/src/main/groovy/uk/ac/ox/softeng/mauro/api/datamodel/DataModelApi.groovy
@@ -9,15 +9,14 @@ import uk.ac.ox.softeng.mauro.api.model.ModelVersionedRefDTO
 import uk.ac.ox.softeng.mauro.api.model.ModelVersionedWithTargetsRefDTO
 import uk.ac.ox.softeng.mauro.api.model.VersionLinkDTO
 import uk.ac.ox.softeng.mauro.domain.datamodel.DataModel
-import uk.ac.ox.softeng.mauro.domain.datamodel.DataModelType
 import uk.ac.ox.softeng.mauro.domain.datamodel.IntersectsData
 import uk.ac.ox.softeng.mauro.domain.datamodel.IntersectsManyData
 import uk.ac.ox.softeng.mauro.domain.datamodel.SubsetData
 import uk.ac.ox.softeng.mauro.domain.diff.ObjectDiff
 import uk.ac.ox.softeng.mauro.domain.model.version.CreateNewVersionData
 import uk.ac.ox.softeng.mauro.domain.model.version.FinaliseData
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchRequestDTO
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchResultsDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchRequestDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchResultsDTO
 import uk.ac.ox.softeng.mauro.plugin.datatype.DataTypePlugin
 import uk.ac.ox.softeng.mauro.plugin.importer.DataModelImporterPlugin
 import uk.ac.ox.softeng.mauro.web.ListResponse
