+++ mauro-client/src/main/groovy/uk/ac/ox/softeng/mauro/api/search/SearchApi.groovy
@@ -2,15 +2,14 @@ package uk.ac.ox.softeng.mauro.api.search
 
 import uk.ac.ox.softeng.mauro.api.MauroApi
 import uk.ac.ox.softeng.mauro.api.Paths
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchRequestDTO
-import uk.ac.ox.softeng.mauro.persistence.search.dto.SearchResultsDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchRequestDTO
+import uk.ac.ox.softeng.mauro.domain.search.dto.SearchResultsDTO
 import uk.ac.ox.softeng.mauro.web.ListResponse
 
 import io.micronaut.context.annotation.Parameter
 import io.micronaut.http.annotation.Body
 import io.micronaut.http.annotation.Get
 import io.micronaut.http.annotation.Post
-import io.micronaut.http.annotation.RequestBean
 
 @MauroApi
 interface SearchApi {
